Add test code for SoundHDF5File
